Restructure item collection
Firstly, change "item" -> "entity" since gold isn't an item,
but this function does apply for gold.

Next, the function itself just checks the entity type, and emits
different signals based on if the entity is of GOLD type or not.

The function also emits the entity itself when it is an item type,
since it is now the player's responsibility to deal with the item node
when interacting with the item.

This responsibility change is because the item doesn't (and shouldn't)
know if the player already has a copy of it. Thus, the player must check
if it already has one and decide whether to hide or leave the item on
the floor. However, the GOLD type doesn't need to worry about that,
since the player can't pick up the gold entity regardless.