Add item stats scripts
Each of these scripts gets the name of the entity it's associated with,
and depending on what the name is, changes its stat based on it. This
makes future balancing easy, since all armor stats are just in
armor_stats. This happens when any of these objects are created, so they
will automatically have the correct defense, attack, potion effect, etc.
on instantiation.

Side note: for now, since keys have no real use, the misc_stats file
just says that the stat is that it opens something. This is meaningless,
but it is just a placeholder for future behavior.