Add script to classify entities
This script simply checks the item name and classifies the item into
types based on the name. It uses an enum to represent types. This enum
is also present in other files, since I was having issues referencing it
in other scripts, and online research mentioned it might be a bug with
Godot 4.0. I might make it global in the future, but for now, just copy
pasting one line isn't too bad.