Add inventory node to Player
Currently, this is unimplemented, but in the future, if we want to drop items,
instead of deleting items from the scene, we can simply hide them, and
link them to the player's node. Then, when we want to drop them, we'll
just unparent them from the player and make them visible.

For now, we just add and hide items to be picked up to this inventory
node. Dropping is unimplemented, but this provides a good framework for
the future.

This inventory node will add items in the same order that the string
inventory in player does, so they will be synced at all times.