Modify how players pick up items
Extra inventory referencing Inventory node added, so that we can access
both the instantiated hidden nodes in our inventory, but also keep the
string inventory for quick checking of whether we have an item.

For the function changes, we now just check if we have the item based on
the item name, and if we don't have it, we add it to both inventories,
but if we do have it, we print that we have one and leave the item on
the ground. This completes the unique picking up of items.

In the future, I'll add extra checks for stackable items such as arrows,
but since those are unimplemented, I have no need to add that yet.

This change does remove the functionality of items increasing stats, but
the stats are still present in the item nodes themselves. This is the
start of the feature of using items. Before this change, you could
infinitely stack sword attacks, getting insanely high attack and
defense. However, when I get around to wielding weapons and wearing
armor, the functionality of the items will be restored.