Modify collect script to only delete GOLD types
The function has been changed to only call queue_free() if the entity is
of the GOLD type.

All pickups will send the whole node to GameMaster, so that it can send
the node to Player if the entity is an item. Once at the player, the
player can decide whether to claim or leave the item on the ground.

This also removes the unused and deprecated _on_area_entered call.