Player uses most recently picked up item
As a temporary fix to allow the player's stats to change when picking up
items, the player's strength and defense scores are modified to reflect
the most recently picked up item. This can result in strength and
defense scores going down if the player picks up a worse item, but is
not meant to be a permanent change, just until I can get using items
implemented.

Also added base strength/defense scores for these
calculations.