actually fix shader issue
the reason the psychedelic potion and darkness scroll were doing the
same shader was because they were using the same material. So when I
"fixed it" by setting one to a different shader, I was actually just
switching the shader for that material, but both ColorRects were using
the same material, so it would always swap both of them. Made another
material to fix this issue, so they both have individual correct
behavior now.